backend/pyproject.toml
[project]
name = "predoc-backend"
version = "0.1.0"
description = "AI Pre-consultation Agent based on TCM knowledge base"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "faiss-cpu>=1.8.0",
    "unstructured>=0.13.0",
    "python-docx>=1.1.0",
    "pydantic>=2.6.0",
    "sse-starlette>=2.0.0",
    "aiofiles>=23.2.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
